Bugfixes for IPSEC, and a fixed warning in MikeyMessage
git-svn-id: svn://svn.minisip.org/minisip/trunk@1052 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libmikey/include/libmikey/keyagreement.h
@@ -49,12 +49,12 @@
 // Class to hold Security Policy (SP) info
 class Policy_type {
 	public:
-		Policy_type(uint8_t policy_No, uint8_t prot_type, uint8_t policy_type, uint16_t length, byte_t * value);
+		Policy_type(uint8_t policy_No, uint8_t prot_type, uint8_t policy_type, uint8_t length, byte_t * value);
 		~Policy_type();
 		uint8_t policy_No;
 		uint8_t prot_type;
 		uint8_t policy_type;
-		uint16_t length;
+		uint8_t length;
 		byte_t * value;
 	private:
 };
